* Update
git-svn-id: https://svn.code.sf.net/p/gmerlin/code/trunk/gmerlin_avdecoder@5251 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ lib/parse_opus.c
@@ -53,7 +53,7 @@ static int get_format(bgav_audio_parser_t * parser)
     bgav_opus_header_dump(&h);
 
   parser->s->ci.pre_skip = h.pre_skip;
-  p->pts_offset = -parser->s->ci.pre_skip;
+  p->pts_offset = - ((int)parser->s->ci.pre_skip);
   ret = 1;
   fail:
